Jgonzales/bs 123 update level (#32)
* Update level descriptions

* Squashed 'entities/' changes from 0b79648..9dfd906

9dfd906 Entities for studens level
1a76111 Merge branch 'main' into jgonzales/BS-123_student_level
dbed9e0 Fix minor errors
cd2a488 Fix type for operation number
2784cd3 Merge commit 'b37a018af018b195d7da85568d7de3614afddb32' into nypex10/BS-120_validate-operation
897270f Merge commit '64e91418e52254bc9f0eba21ab66b30adeab4346' into jgonzales/BS-129_add_operation_number
1d52d88 Add entities for add operation number
274d094 Merge branch 'main' into nypex10/BS-120_validate-operation
1d01d36 Fix (#404)
a537270 Change game (#403)
70cf847 Merge commit 'e5f8e059bd02050c191bb5a86bed98370075b96c' into jgonzales/BS-124_moods
3653f2c Entities for moods
30987d4 Update level from harcode to table in database
a61fa13 Update level descriptions (#130)
532a15c Update levels description (#401)
a20e552 Merge commit '4840cdb3000f09216d07ec8eb27ab9d59265c55f' into jgonzales/BS-121_operation_number_services
f52f502 Added entities for operation number
38c2fa3 Merge commit '21a6e84e111ad03a43365125d860ee18e3b1b2f7' into jgonzales/BS-103_control_report
fce1120 Add car control report (#393)
e50197f Add car control report
a0ab65b Jgonzales/bs 98 add car control (#391)
a0bd6bf Merge commit '0355b620bdecbd7f7db52a927f0a6044216d5612' into jgonzales/BS-98_add_car_control
6306ee4 Entities for list, add and remove license plates for car control
c14e224 Jgonzales/bs 96 add license plate (#390)
d92cb5c Fix update service
e63599e merge
cf77d85 Added license plate to student services
ae0071e Jgonzales/bs 87 update socios report (#384)
259ec96 Remove source
e7f1b19 Merge commit '3f9d2af1e6dd35e34e8f4b1cf73d41305ddf63fc' into jgonzales/BS-87_update_socios_report
c2c4547 Entities for report pacasmayo socio students
f66d2a7 Jgonzales/bs 77 report socios pacasmayo (#372)
458cde8 Merge commit '2c68fc44137af492fc15616d44b31c5fc6f5948d' into jgonzales/BS-77_report_socios_pacasmayo
30d3c3a Report socio pacasmayo parameters
c5df07b Jgonzales/bs 75 set pool sut (#371)
ca04d55 Add poolId as paramater
16568fb Merge commit '9d353f940550e991bcd8b471edeac658f37b054b' into jgonzales/BS-75_set_pool_sut
6e6ef04 Jgonzales/bs 68 sut products (#364)
01b9e7d Update sale params
dad3967 Merge main entities
ad7d0f6 update list product service to search by variation type, value and location
9a34d54 Entities to delete product, variation and stock
0518960 Stock add and update params dto
c789de1 Make the variationTypeId optional
a0340f3 Merge commit 'e13e0959efe131ccf96c3cdc22f69b9d800004ef' into jgonzales/BS-68_sut_products
5b922b0 Add variation add and update params dto
6ab2ae2 Update
54d053d Merge commit 'ca3e103ed28ae3c3e4e1dbaa893b471d5441120f' into jgonzales/BS-68_sut_products
6130991 Product update
d84c9d6 Merge commit 'a3bd4a414c7e89f093ad0bae7bccbd65572e024b' into jgonzales/BS-68_sut_products
3188e9b remove unused libraries
3490c6a Product edit
5cbd4c6 Merge commit '49d5b813484a988a098d3eef020c76c41f8abfce' into jgonzales/BS-68_sut_products
4b4ef84 Merge into main
ce18eb8 New entities
062bc93 Jgonzales/bs 61 club inactive users (#357)
2d267fb Jgonzales/bs 61 club inactive users (#103)
98ef1dc Merge commit '87fb72c932b5b3dd348f7a1a623f5c287c27ea4e' into jgonzales/BS-61_club_inactive_users
da1b3c3 Add is active in list club member service
c7cbfa9 Jgonzales/bs 60 sut report (#208)
3d90cd5 Jgonzales/bs 60 sut report (#102)
deb9b4e Jgonzales/bs 60 sut report (#356)
eaee489 Fix bug
e1c9000 Add new pool parameter
66a3c56 Merge commit 'ad6a390b4068cd55b83b116fdd897e577ac7bb03' into jgonzales/BS-60_sut_report
0824817 Jgonzales/bs 58 fixes (#101)
c771682 Jgonzales/bs 58 fixes (#354)
dd8a55f change survey_pool_id by pool_id
65860d6 merge entities
b8bd98a update survey results
a3d0cf6 Jgonzales/bs 59 sut (#207)
bcf754e Merge entities
ec8023f Jgonzales/bs 57 survey update (#352)
387a67b Merge main entities
ecced06 Merge main entities
f93fb73 Jgonzales/bs 57 survey update (#99)
192723d Add new survey
2c23388 New survey
17e1ed3 Dev (#350)
07fb9b8 Merge commit '6f53295d19bf3459f4916a7dd73e0445cd7e37b5' into dev
6c06cc8 Merge branch 'main' into dev
237dca5 Jgonzales/bs 54 preequipo asistencia (#346)
5ece5b9 Jgonzales/bs 52 maintenance (#204)
8afbf60 Jgonzales/bs 52 maintenance (#200)
e5d3e1b Jgonzales/bs 52 maintenance (#199)

git-subtree-dir: entities
git-subtree-split: 9dfd9064207dbc304391f687e1f41f8e2c07863d

* Student level from database

* Squashed 'entities/' changes from 9dfd906..5ad9792

5ad9792 Replace level type and validation to integer
a2c1f30 Replace level type and validation to integer
bb6d70f Merge entities
b1c1777 Update level descriptions (#226)
29f5d5f Jgonzales/bs 96 add license plate (#222)
185ba32 Jgonzales/bs 75 set pool sut (#215)
0e7c179 Jgonzales/bs 68 sut products (#212)

git-subtree-dir: entities
git-subtree-split: 5ad9792db84458cb5738cf15fee3f4eee32a0cb8

* Squashed 'entities/' changes from 5ad9792..d442275

d442275 Remove unused type
06eac65 Merge commit '4ca0e0b4fd743758786b2393b63903d440592e63' into jgonzales/BS-123_student_level
ab90bca Add new sellers (#133)

git-subtree-dir: entities
git-subtree-split: d4422757691b524483a17a4d89317e7ce60cc925

---------

+++ components/auth/auth.tsx
@@ -123,6 +123,10 @@ const AlAuth: React.FC<AlAuthProps> = ({ children, auth }) => {
         context.user,
     ]);
 
+    if (route === "idle") {
+        return <></>;
+    }
+
     if (route === "signIn") {
         if (typeof window !== "undefined") {
             window.localStorage.clear();
